+++ internal/applets/shellutils/dirname/dirname.go
@@ -1,0 +1,100 @@
+//
+// mimixbox/internal/applets/shellutils/dirname/dirname.go
+//
+// Copyright 2021 Naohiro CHIKAMATSU
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package dirname
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"github.com/jessevdk/go-flags"
+	mb "github.com/nao1215/mimixbox/internal/lib"
+)
+
+const cmdName string = "dirname"
+
+const version = "1.0.0"
+
+var osExit = os.Exit
+
+// Exit code
+const (
+	ExitSuccess int = iota // 0
+	ExitFailuer
+)
+
+type options struct {
+	Version bool `short:"v" long:"version" description:"Show dirname command version"`
+	Zero    bool `short:"z" long:"zero" description:"Print each output line without line feed"`
+}
+
+func Run() (int, error) {
+	var opts options
+	var args []string
+	var err error
+
+	if args, err = parseArgs(&opts); err != nil {
+		return ExitFailuer, nil
+	}
+
+	return dirname(args, opts)
+}
+
+func dirname(args []string, opts options) (int, error) {
+	status := ExitSuccess
+	for _, path := range args {
+		dirname := filepath.Dir(os.ExpandEnv(path))
+		if opts.Zero {
+			fmt.Fprintf(os.Stdout, "%s", dirname)
+		} else {
+			fmt.Fprintln(os.Stdout, dirname)
+		}
+	}
+	return status, nil
+}
+
+func parseArgs(opts *options) ([]string, error) {
+	p := initParser(opts)
+
+	args, err := p.Parse()
+	if err != nil {
+		return nil, err
+	}
+
+	if opts.Version {
+		mb.ShowVersion(cmdName, version)
+		osExit(ExitSuccess)
+	}
+
+	if !isValidArgNr(args) {
+		fmt.Fprintln(os.Stderr, "dirname: no operand")
+		osExit(ExitFailuer)
+	}
+	return args, nil
+}
+
+func initParser(opts *options) *flags.Parser {
+	parser := flags.NewParser(opts, flags.Default)
+	parser.Name = cmdName
+	parser.Usage = "[OPTIONS] PATH"
+
+	return parser
+}
+
+func isValidArgNr(args []string) bool {
+	return len(args) >= 1
+}